@@ -5,7 +5,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: '>=3.5.2 <4.0.0'
 
 dependencies:
   flutter:
@@ -25,11 +25,11 @@ dependencies:
   firebase_auth: ^4.9.0
   google_sign_in: ^6.1.5
   get_it: ^7.6.2
-  flutter_native_splash: ^2.3.2
+  flutter_native_splash: ^2.4.1
   mocktail: ^1.0.0
   email_validator: ^2.1.17
   dio: ^5.3.3
-  flutter_carousel_widget: ^2.1.2
+  flutter_carousel_widget: ^2.3.0
   image_picker: ^1.0.4
   flutter_staggered_grid_view: ^0.7.0
   http_parser: ^4.0.2
@@ -40,12 +40,19 @@ dependencies:
   flutter_quill:
     git: 
       url: https://github.com/cledsonv/flutter-quill.git
-  vsc_quill_delta_to_html: ^1.0.4
-  flutter_widget_from_html_core: ^0.14.11
+  vsc_quill_delta_to_html: ^1.0.5
+  flutter_widget_from_html_core: ^0.15.2
   dotted_border: ^2.1.0
   infinite_scroll_pagination: ^4.0.0
+  page_manager:
+    git:
+      url: https://github.com/lucascardo12/page_manager.git
+
 dependency_overrides:
   intl: ^0.19.0
+  win32: ^5.5.4
+  archive: ^3.6.1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -64,4 +71,4 @@ flutter_native_splash:
   color: "#FFFFFF"
   android_12:
     image: "assets/images/logo.png"
-    icon_background_color: "#101636"
+    icon_background_color: "#FFFFFF"
