@@ -1,7 +1,9 @@
 import 'package:app_receitas/src/feactures/auth/domain/services/auth_serivce.dart';
+import 'package:app_receitas/src/feactures/home/presenter/ui/pages/custom_bottom_bar.dart';
 import 'package:flutter/material.dart';
+import 'package:page_manager/export_manager.dart';
 
-class AuthController extends ChangeNotifier {
+class AuthController extends ManagerStore {
   final AuthService _authService;
 
   AuthController(
@@ -12,36 +14,53 @@ class AuthController extends ChangeNotifier {
   TextEditingController passwordController = TextEditingController();
 
   @override
+  void init(Map<String, dynamic> arguments) {
+    state = StateManager.done;
+  }
+
+  @override
   void dispose() {
     super.dispose();
     emailController.dispose();
     passwordController.dispose();
   }
 
-  Future<bool> registerFirebase() async {
-    await _authService.signUp(
-      email: emailController.text.trim(),
-      password: passwordController.text.trim(),
-    );
-    return true;
-  }
+  Future<void> registerFirebase() => handleTry(
+        call: () async {
+          await _authService.signUp(
+            email: emailController.text.trim(),
+            password: passwordController.text.trim(),
+          );
+          emitNavigation(CustomBottomBar.route);
+        },
+        showDialogError: true,
+      );
 
-  Future<bool> loginFirebase() async {
-    await _authService.signIn(
-      email: emailController.text.trim(),
-      password: passwordController.text.trim(),
-    );
-    return true;
-  }
+  Future<void> loginFirebase() => handleTry(
+        call: () async {
+          await _authService.signIn(
+            email: emailController.text.trim(),
+            password: passwordController.text.trim(),
+          );
 
-  Future<void> forgetPassword() async {
-    await _authService.forgetPassword(
-      email: emailController.text.trim(),
-    );
-  }
+          emitNavigation(CustomBottomBar.route);
+        },
+        showDialogError: true,
+      );
 
-  Future<bool> loginGoogle() async {
-    await _authService.signInGoogle();
-    return true;
-  }
+  Future<void> forgetPassword() => handleTry(
+        call: () async {
+          await _authService.forgetPassword(
+            email: emailController.text.trim(),
+          );
+        },
+        showDialogError: true,
+      );
+
+  Future<void> loginGoogle() => handleTry(
+        call: () async {
+          await _authService.signInGoogle();
+          emitNavigation(CustomBottomBar.route);
+        },
+      );
 }
