@@ -1,14 +1,13 @@
-import 'package:app_receitas/src/core/global/global_variables.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_button.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_text_field.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_text_span.dart';
 import 'package:app_receitas/src/feactures/auth/presenter/controllers/auth_controller.dart';
 import 'package:app_receitas/src/feactures/auth/presenter/ui/organisms/custom_screen.dart';
-import 'package:app_receitas/src/feactures/home/presenter/ui/pages/custom_bottom_bar.dart';
 import 'package:email_validator/email_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:page_manager/export_manager.dart';
 
 class RegisterPage extends StatefulWidget {
   static const route = '/register';
@@ -18,21 +17,23 @@ class RegisterPage extends StatefulWidget {
   State<RegisterPage> createState() => _RegisterPageState();
 }
 
-class _RegisterPageState extends State<RegisterPage> {
-  AuthController ct = di();
+class _RegisterPageState extends ManagerPage<AuthController, RegisterPage> {
   final formKey = GlobalKey<FormState>();
-
   @override
-  void dispose() {
-    ct.dispose();
-    super.dispose();
+  void initState() {
+    ct.onNavigation(
+      (event) => Navigator.pushNamedAndRemoveUntil(
+        context,
+        event,
+        (route) => false,
+      ),
+    );
+    super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
     ColorScheme color = Theme.of(context).colorScheme;
-    final theme = Theme.of(context).colorScheme;
-    final snack = ScaffoldMessenger.of(context);
     return CustomScreen(
       iconAppBar: IconButton(
         onPressed: () => Navigator.pop(context),
@@ -97,27 +98,7 @@ class _RegisterPageState extends State<RegisterPage> {
             CookieButton(
               label: 'Google',
               onPressed: () async {
-                final result = await ct.loginGoogle();
-                if (!result && context.mounted) {
-                  final snackBar = SnackBar(
-                    content: CookieText(
-                      text: AppLocalizations.of(context)!.registerTextSnack,
-                      color: theme.onSecondary,
-                    ),
-                    action: SnackBarAction(
-                      label: AppLocalizations.of(context)!.registerOptionSnack,
-                      onPressed: () {},
-                    ),
-                  );
-                  snack.showSnackBar(snackBar);
-                }
-                if (result && context.mounted) {
-                  Navigator.pushNamedAndRemoveUntil(
-                    context,
-                    CustomBottomBar.route,
-                    (route) => false,
-                  );
-                }
+                await ct.loginGoogle();
               },
             ),
             const SizedBox(height: 10),
@@ -154,28 +135,7 @@ class _RegisterPageState extends State<RegisterPage> {
           label: AppLocalizations.of(context)!.registerButton,
           onPressed: () async {
             if (formKey.currentState!.validate()) {
-              var result = await ct.registerFirebase();
-
-              if (!result && context.mounted) {
-                final snackBar = SnackBar(
-                  content: CookieText(
-                    text: AppLocalizations.of(context)!.registerTextSnack,
-                    color: theme.onSecondary,
-                  ),
-                  action: SnackBarAction(
-                    label: AppLocalizations.of(context)!.registerOptionSnack,
-                    onPressed: () {},
-                  ),
-                );
-                snack.showSnackBar(snackBar);
-              }
-              if (result && context.mounted) {
-                Navigator.pushNamedAndRemoveUntil(
-                  context,
-                  CustomBottomBar.route,
-                  (route) => false,
-                );
-              }
+              await ct.registerFirebase();
             }
           },
         ),
