@@ -3,12 +3,14 @@ import 'package:app_receitas/src/core/widgets/feactures/cookie_page.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
 import 'package:app_receitas/src/core/widgets/feactures/cookie_text_field_search.dart';
 import 'package:app_receitas/src/feactures/home/presenter/controller/home_controller.dart';
+import 'package:app_receitas/src/feactures/onboarding/presenter/ui/pages/onboarding_page.dart';
 import 'package:app_receitas/src/feactures/profile/presenter/ui/atomic/container_profile_image.dart';
 import 'package:app_receitas/src/feactures/recipes/presenter/ui/pages/create_recipe_page.dart';
 import 'package:app_receitas/src/feactures/search/presenter/ui/pages/search_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:page_manager/manager_page.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -17,15 +19,15 @@ class HomePage extends StatefulWidget {
   State<HomePage> createState() => _HomePageState();
 }
 
-class _HomePageState extends State<HomePage> {
-  final HomeController ct = di();
-
+class _HomePageState extends ManagerPage<HomeController, HomePage> {
   @override
   void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((_) => ct.init(context));
-    ct.addListener(() {
-      setState(() {});
-    });
+    WidgetsBinding.instance.addPostFrameCallback(
+      (_) {
+        ct.verifyOnboading().then((value) =>
+            value ? Navigator.pushNamed(context, OnBoardingPage.route) : null);
+      },
+    );
     super.initState();
   }
 
@@ -40,7 +42,7 @@ class _HomePageState extends State<HomePage> {
         backgroundColor: Theme.of(context).colorScheme.primary,
         child: SvgPicture.asset('assets/icons/edit.svg'),
       ),
-      done: (_) => SafeArea(
+      done: () => SafeArea(
         child: Padding(
           padding: const EdgeInsets.fromLTRB(10, 10, 10, 5),
           child: ListView(
