@@ -1,12 +1,11 @@
 import 'dart:io';
-
-import 'package:app_receitas/src/core/widgets/feactures/cookie_page.dart';
 import 'package:app_receitas/src/feactures/profile/domain/entities/profile_entity.dart';
 import 'package:app_receitas/src/feactures/profile/domain/repositories/profile_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:page_manager/export_manager.dart';
 
-class EditProfileController extends ChangeNotifier {
+class EditProfileController extends ManagerStore {
   final ProfileRepository _repository;
 
   EditProfileController(this._repository);
@@ -14,17 +13,15 @@ class EditProfileController extends ChangeNotifier {
   File? image;
   ProfileEntity? profile;
   TextEditingController biographyController = TextEditingController();
-  PageState state = PageState.loading;
   bool isRemoveImage = false;
 
-  void init(context) async {
-    state = PageState.loading;
-    final arguments = ModalRoute.of(context)!.settings.arguments;
-    profile = arguments as ProfileEntity;
-    biographyController.text = profile!.biography;
-    state = PageState.done;
-    notifyListeners();
-  }
+  @override
+  void init(Map<String, dynamic> arguments) => handleTry(
+        call: () async {
+          profile = arguments['profile'] as ProfileEntity;
+          biographyController.text = profile!.biography;
+        },
+      );
 
   Future<void> updateImageProfile() async {
     if (image == null) return;
